Add files via upload
Build a Temperature Conversion Program.

Overview

A programm that converts temperature between Celsious, Fahrenheit, and Kelvin scales. In program users to input a temperature value and the original unit of measurement.

Technologies Used

C programming

Acknowledgments Thanks to Prodigy Infotech for the opportunity and support during this internship. Special thanks to my mentors and colleagues for their guidance.